pyproject.toml
[project]
name = "libraryapi"
version = "0.1.0"
description = ""
authors = [
    {name = "yuraaceventura",email = "yuraaceventura@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
